refactor: rename takePicture to takePhoto
