+++ app/src/main/java/be/hogent/faith/faith/di/KoinModules.kt
@@ -4,7 +4,7 @@ import be.hogent.faith.faith.createUser.CreateEventViewModel
 import be.hogent.faith.faith.mainScreen.MainScreenViewModel
 import be.hogent.faith.faith.drawEmotionAvatar.DrawEmotionViewModel
 import be.hogent.faith.faith.enterEventDetails.EventDetailsViewModel
-import be.hogent.faith.faith.takePicture.TakePhotoViewModel
+import be.hogent.faith.faith.takePhoto.TakePhotoViewModel
 import io.reactivex.android.schedulers.AndroidSchedulers
 import org.koin.android.viewmodel.ext.koin.viewModel
 import org.koin.dsl.module.module
