+++ app/src/main/res/layout/fragment_take_photo.xml
@@ -5,7 +5,7 @@
 
         <variable
                 name="photoViewModel"
-                type="be.hogent.faith.faith.takePicture.TakePhotoViewModel"/>
+                type="be.hogent.faith.faith.takePhoto.TakePhotoViewModel"/>
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
